Added show/hidden functionality with dropdown menu. Moved accordion from Navbar to NavbarAccordion component. Updated styling to maintain spatial continuity.
