+++ travelers-compendium/src/App.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Navbar from './components/navbar/Navbar.jsx';
 import Body from './components/Body/Body.jsx';
 import Footer from './components/Footer/Footer.jsx';
