+++ travelers-compendium/src/styles/Body.css
@@ -1,0 +1,4 @@
+.body {
+    /* position: absolute;
+    z-index: 1; */
+}
