+++ travelers-compendium/src/styles/Navbar.css
@@ -2,19 +2,13 @@
 @import url('https://fonts.googleapis.com/css2?family=Piazzolla&display=swap');
 
 .mainNavBar {
-    display: inline-block;
-    /* background-color: rgb(27, 220, 117); */
+    display: flex;
+    flex-direction: column;
     background-image: url('../../public/Header.png');
     width: 100%;
-    /* height: 22vh; */
     background-size: 100%;
     background-repeat: no-repeat;
-    /* border: 2px solid black; */
-    /* border-radius: 10px; */
-    font-size: 1.25em;
     margin-bottom: 2.5vh;
-    align-items: center;
-    justify-content: center;
 }
 
 .mainTitle {
@@ -27,8 +21,10 @@
 .navBarLogo {
     float: left;
     width: 12%;
+    height: 12%;
     padding: 30px;
     cursor: pointer;
+    z-index: 5;
 }
 
 .navBarLogo:hover {
@@ -37,11 +33,15 @@
 }
 
 .navButtonContainer {
+    position: fixed;
+    top: 16%;
     box-sizing: inherit;
-    width: 20%;
-    height: 100vh;
+    width: 15%;
+    /* height: 100vh; */
     display: block;
     float: left;
+    visibility: hidden;
+    z-index: 5;
 }
 
 .navButtonContainer p {
